:recycle: Split huge logic test file
